requirements.txt
streamlit>=1.28.0
langchain>=0.3.27
langchain-perplexity==0.1.2
python-dotenv>=1.0.1
twilio>=8.9.0
requests>=2.32.0
